14. 4. 17
- Change checking method about guild members which using KnightFrame
Armory function.
- Addon will check him when you popup the unit menu by right clicking
guild member in the guild list.
- I had avoided this way because guild popup menu will always have a
inspect order in this way but past method had a defect that addon send a
addon message in guild channel immediatly when I log on but in this
situation, guild channel was not synchronized other guild members (few
seconds after log on, maybe 2~3 sec but this lag was always different.
shit) so other guild member couldn't take a addon message.